refactor(auth): relocated login and register pages into auth folder
