+++ client/src/pages/auth/RegisterPage.jsx
